refactor: 상품 바텀시트 내부 컴포넌트 네이밍 변경
