+++ src/pages/home/components/product/SelectedProductSheet/SelectedProductSheet.css.ts
